Use one thread for NIO
Fixes #11
